[RF] More `RooNaNPacker` to RooBatchCompute
So we can eventually also use it on the GPU.
